optimized index.html
Modified index.html:
Moved Google Analytics to async linked JS script.
Fixed the reference to Google Fonts that brought in every version and
language of the font.
Fixed the CSS, removed extraneous and redundant CSS styles.
Made local copy of the linked images because the response included it's
own DOM! (The response was an image served in a DOM.)