Improvements to Pizza page
Tweaked settings, reviewed assignments to try and find ways to squeeze
more fps.  I think we are ready for production.  Anything more at this
point is more work than necessary.  Speeds surpass the expected and that
is good enough for me.  Now to run a build, put it in production, test
and document everything!